dot
Enum ArrowShape

pub enum ArrowShape {
    NoArrow,
    Normal(Fill, Side),
    Box(Fill, Side),
    Crow(Side),
    Curve(Side),
    ICurve(Fill, Side),
    Diamond(Fill, Side),
    Dot(Fill),
    Inv(Fill, Side),
    Tee(Side),
    Vee(Side),
}

This enumeration represents all possible arrow edge as defined in graphviz documentation.

Variants

NoArrow

No arrow will be displayed

Normal(Fill, Side)

Arrow that ends in a triangle. Basically a normal arrow. NOTE: there is error in official documentation, this supports both fill and side clipping

Box(Fill, Side)

Arrow ending in a small square box

Crow(Side)

Arrow ending in a three branching lines also called crowโs foot

Curve(Side)

Arrow ending in a curve

ICurve(Fill, Side)

Arrow ending in an inverted curve

Diamond(Fill, Side)

Arrow ending in an diamond shaped rectangular shape.

Dot(Fill)

Arrow ending in a circle.

Inv(Fill, Side)

Arrow ending in an inverted triangle.

Tee(Side)

Arrow ending with a T shaped arrow.

Vee(Side)

Arrow ending with a V shaped arrow.

Implementations

impl ArrowShape

pub fn none() -> ArrowShape

Constructor which returns no arrow.

pub fn normal() -> ArrowShape

Constructor which returns normal arrow.

pub fn boxed() -> ArrowShape

Constructor which returns a regular box arrow.

pub fn crow() -> ArrowShape

Constructor which returns a regular crow arrow.

pub fn curve() -> ArrowShape

Constructor which returns a regular curve arrow.

pub fn icurve() -> ArrowShape

Constructor which returns an inverted curve arrow.

pub fn diamond() -> ArrowShape

Constructor which returns a diamond arrow.

pub fn dot() -> ArrowShape

Constructor which returns a circle shaped arrow.

pub fn inv() -> ArrowShape

Constructor which returns an inverted triangle arrow.

pub fn tee() -> ArrowShape

Constructor which returns a T shaped arrow.

pub fn vee() -> ArrowShape

Constructor which returns a V shaped arrow.

pub fn to_dot_string(&self) -> String

Function which renders given ArrowShape into a String for displaying.