Enum dot::ArrowShape

pub enum ArrowShape {
    NoArrow,
    Normal(Fill, Side),
    Box(Fill, Side),
    Crow(Side),
    Curve(Side),
    ICurve(Fill, Side),
    Diamond(Fill, Side),
    Dot(Fill),
    Inv(Fill, Side),
    Tee(Side),
    Vee(Side),
}

This enumeration represents all possible arrow edge as defined in grapviz documentation.

Variants

NoArrow

No arrow will be displayed

Normal

Arrow that ends in a triangle. Basically a normal arrow. NOTE: there is error in official documentation, this supports both fill and side clipping

Box

Arrow ending in a small square box

Crow

Arrow ending in a three branching lines also called crow's foot

Curve

Arrow ending in a curve

ICurve

Arrow ending in an inverted curve

Diamond

Arrow ending in an diamond shaped rectangular shape.

Dot

Arrow ending in a circle.

Inv

Arrow ending in an inverted triangle.

Tee

Arrow ending with a T shaped arrow.

Vee

Arrow ending with a V shaped arrow.

Methods

impl ArrowShape

fn none() -> ArrowShape

Constructor which returns no arrow.

fn normal() -> ArrowShape

Constructor which returns normal arrow.

fn boxed() -> ArrowShape

Constructor which returns a regular box arrow.

fn crow() -> ArrowShape

Constructor which returns a regular crow arrow.

fn curve() -> ArrowShape

Constructor which returns a regular curve arrow.

fn icurve() -> ArrowShape

Constructor which returns an inverted curve arrow.

fn diamond() -> ArrowShape

Constructor which returns a diamond arrow.

fn dot() -> ArrowShape

Constructor which returns a circle shaped arrow.

fn inv() -> ArrowShape

Constructor which returns an inverted triangle arrow.

fn tee() -> ArrowShape

Constructor which returns a T shaped arrow.

fn vee() -> ArrowShape

Constructor which returns a V shaped arrow.

fn to_dot_string(&self) -> String

Function which renders given ArrowShape into a String for displaying.