Struct dot::Arrow

pub struct Arrow {
    pub arrows: Vec<ArrowShape>,
}

This structure holds all information that can describe an arrow connected to either start or end of an edge.

Fields

arrows

Methods

impl Arrow

fn none() -> Arrow

Arrow constructor which returns an empty arrow

fn normal() -> Arrow

Arrow constructor which returns a regular triangle arrow, without modifiers

fn from_arrow(arrow: ArrowShape) -> Arrow

Arrow constructor which returns an arrow created by a given ArrowShape.

fn to_dot_string(&self) -> String

Function which converts given arrow into a renderable form.